requirements.txt
asgiref==3.7.2
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
pytz==2024.1
sqlparse==0.4.4
tzdata==2023.4
